requirements.txt
numpy==1.24.4
faiss-cpu==1.7.4
Flask==2.3.3
sentence-transformers==2.2.2